requirements.txt
crawl4ai==0.6.3
python-dotenv
click
rich 